+++ client/components/business/NavBusiness.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Container, Button, Menu } from 'semantic-ui-react'
-import { logOut } from '../../api/passportAPI'
+import { logOut } from '../../api/apiClient'
 import { Link } from 'react-router-dom'
 import { businessInfo } from '../Login'
