+++ client/components/customer/NavCustomer.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Container, Button, Menu, Image, Icon } from 'semantic-ui-react'
 import { Link } from 'react-router-dom'
-import { logOut } from '../../api/passportAPI'
+import { logOut } from '../../api/apiClient'
 import { userInfo } from '../Login'
 
 export default function NavCustomer () {
